Add session control, clean code

+++ server/app.js
@@ -11,6 +11,17 @@ const User = require("./model/user");
 const Contact = require("./model/contact");
 
 const url = "mongodb://0.0.0.0:27017/ContactsDB";
+const statusCode = {
+  "Bad Request": 400,
+  Error: 400,
+  Unauthorized: 401,
+  Forbidden: 403,
+  "Not Found": 404,
+  Conflict: 409,
+  "Username Conflict": 409,
+  "Email Conflict": 409,
+  "Phone number Conflict": 409,
+};
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
@@ -29,55 +40,43 @@ app.post("/api/signup", async (req, res) => {
       phoneNumber: req.body.phoneNumber,
     });
 
-    if (username || email || phoneNumber) {
-      if (username) {
-        return res.status(200).json({
-          status: "success",
-          data: "Username already exist",
-        });
-      } else if (email) {
-        return res.status(200).json({
-          status: "success",
-          data: "Email already exist",
-        });
-      } else {
-        return res.status(200).json({
-          status: "success",
-          data: "Phone number already exist",
-        });
-      }
-    } else {
-      const newUser = new User({
-        username: req.body.username,
-        email: req.body.email,
-        phoneNumber: req.body.phoneNumber,
-        password: req.body.password,
-      });
-
-      bcrypt.genSalt(saltRounds).then(() => {
-        bcrypt
-          .hash(newUser.password, saltRounds)
-          .then((hashedPassword) => {
-            newUser.password = hashedPassword;
-            newUser.save();
-          })
-          .then(() => {
-            return res.status(200).json({
-              status: "success",
-              data: newUser,
-            });
-          })
-
-          .catch((error) => {
-            console.log("Error saving user: ");
-            console.log(error);
-          });
-      });
+    if (username) {
+      throw new Error("Username Conflict");
     }
-  } catch {
-    return res.status(400).json({
+    if (email) {
+      throw new Error("Email Conflict");
+    }
+    if (phoneNumber) {
+      throw new Error("Phone Number Conflict");
+    }
+
+    const newUser = new User({
+      username: req.body.username,
+      email: req.body.email,
+      phoneNumber: req.body.phoneNumber,
+      password: req.body.password,
+      sessionID: "",
+    });
+
+    salt = await bcrypt.genSalt(saltRounds);
+
+    hashedPassword = await bcrypt.hash(newUser.password, saltRounds);
+
+    if (!hashedPassword) {
+      throw new Error("Error");
+    }
+
+    newUser.password = hashedPassword;
+    newUser.save();
+
+    return res.status(200).json({
+      status: "success",
+      data: newUser,
+    });
+  } catch (err) {
+    return res.status(statusCode[err.message]).json({
       status: "fail",
-      data: "Signup error",
+      message: err.message,
     });
   }
 });
@@ -87,6 +86,7 @@ app.post("/api/login", async (req, res) => {
     await mongoose.connect(url);
 
     key = "";
+    sessionID = sessionIDGenerator();
 
     const username = await User.findOne({
       username: req.body.user,
@@ -106,34 +106,54 @@ app.post("/api/login", async (req, res) => {
       key = "phoneNumber";
     }
 
-    User.findOne({ [key]: req.body.user })
-      .then((user) => 
-        bcrypt.compare(req.body.password, user.password)
-      )
-      .then((samePassword) =>
-        !samePassword
-          ? res.status(401).json({
-              status: "fail",
-              data: "Login error",
-            })
-          : res.status(200).json({
-              status: "success",
-              data: "Login",
-            })
-      )
-      .catch((error) => {
-        res.status(401).json({
-          status: "fail",
-          data: "Login error",
-        });
+    user = await User.findOneAndUpdate(
+      { [key]: req.body.user },
+      { sessionID: sessionID }
+    );
 
-        console.log("Error authenticating user: ");
-        console.log(error);
-      });
-  } catch {
-    return res.status(400).json({
+    if (!user) {
+      throw new Error("Not Found");
+    }
+
+    samePassword = await bcrypt.compare(req.body.password, user.password);
+
+    if (!samePassword) {
+      throw new Error("Unauthorized");
+    }
+
+    res.status(200).json({
+      status: "success",
+      data: sessionID,
+    });
+  } catch (err) {
+    return res.status(statusCode[err.message] ?? 400).json({
       status: "fail",
-      data: "Server error",
+      message: err.message,
+    });
+  }
+});
+
+app.post("/api/logout", async (req, res) => {
+  try {
+    await mongoose.connect(url);
+
+    user = await sessionCheck(req.headers.authorization);
+
+    if (!user) {
+      throw new Error("Not Found");
+    }
+
+    user.sessionID = "";
+    user.save();
+
+    res.status(200).json({
+      status: "success",
+      data: "Log out",
+    });
+  } catch (err) {
+    return res.status(statusCode[err.message] ?? 400).json({
+      status: "fail",
+      message: err.message,
     });
   }
 });
@@ -142,6 +162,10 @@ app.get("/api/contacts", async (req, res) => {
   try {
     await mongoose.connect(url);
 
+    if (!(await sessionCheck(req.headers.authorization))) {
+      throw new Error("Unauthorized");
+    }
+
     const contacts = await Contact.find({});
 
     return res.status(200).json({
@@ -149,10 +173,9 @@ app.get("/api/contacts", async (req, res) => {
       data: contacts,
     });
   } catch (err) {
-    return res.status(400).json({
+    return res.status(statusCode[err.message] ?? 400).json({
       status: "fail",
-      data: [],
-      error: err.message,
+      message: err.message,
     });
   }
 });
@@ -161,16 +184,24 @@ app.get("/api/details/:id", async (req, res) => {
   try {
     await mongoose.connect(url);
 
+    if (!(await sessionCheck(req.headers.authorization))) {
+      throw new Error("Unauthorized");
+    }
+
     const contact = await Contact.findById(req.params.id);
+
+    if (!contact) {
+      throw new Error("Not Found");
+    }
 
     return res.status(200).json({
       status: "success",
       data: contact,
     });
   } catch (err) {
-    return res.status(400).json({
+    return res.status(statusCode[err.message] ?? 400).json({
       status: "fail",
-      data: [],
+      message: err.message,
     });
   }
 });
@@ -179,35 +210,43 @@ app.post("/api/add", async (req, res) => {
   try {
     await mongoose.connect(url);
 
-    Contact.findOne({ phoneNumber: req.body.phoneNumber }).then((contact) => {
-      if (contact) {
-        return res.status(200).json({
-          status: "success",
-          data: "Phone number already exist",
-        });
-      } else {
-        const newContact = new Contact({
-          firstName: req.body.firstName,
-          lastName: req.body.lastName,
-          phoneNumber: req.body.phoneNumber,
-          email: req.body.email,
-          addressCity: req.body.addressCity,
-          addressStreet: req.body.addressStreet,
-          isFavorite: req.body.isFavorite,
-        });
+    if (!(await sessionCheck(req.headers.authorization))) {
+      throw new Error("Unauthorized");
+    }
 
-        newContact.save();
-        return res.status(200).json({
-          status: "success",
-          data: newContact,
-        });
-      }
+    existingPhoneNumber = await Contact.findOne({
+      phoneNumber: req.body.phoneNumber,
     });
-  } catch {
-    return res.status(400).json({
+    existingEmail = await Contact.findOne({
+      email: req.body.email,
+    });
+
+    if (existingPhoneNumber) {
+      throw new Error("Phone number Conflict");
+    }
+    if (existingEmail) {
+      throw new Error("Email Conflict");
+    }
+
+    const newContact = new Contact({
+      firstName: req.body.firstName,
+      lastName: req.body.lastName,
+      phoneNumber: req.body.phoneNumber,
+      email: req.body.email,
+      addressCity: req.body.addressCity,
+      addressStreet: req.body.addressStreet,
+      isFavorite: req.body.isFavorite,
+    });
+
+    newContact.save();
+    return res.status(200).json({
+      status: "success",
+      data: newContact,
+    });
+  } catch (err) {
+    return res.status(statusCode[err.message] ?? 400).json({
       status: "fail",
-      data: "Insert error",
-      error: err.message,
+      message: err.message,
     });
   }
 });
@@ -215,6 +254,10 @@ app.post("/api/add", async (req, res) => {
 app.post("/api/edit/:id", async (req, res) => {
   try {
     await mongoose.connect(url);
+
+    if (!(await sessionCheck(req.headers.authorization))) {
+      throw new Error("Unauthorized");
+    }
 
     const filter = { _id: req.params.id };
     const update = {
@@ -231,14 +274,18 @@ app.post("/api/edit/:id", async (req, res) => {
       new: true,
     });
 
+    if (!updatedContact) {
+      throw new Error("Not Found");
+    }
+
     return res.status(200).json({
       status: "success",
       data: updatedContact,
     });
-  } catch {
-    return res.status(400).json({
+  } catch (err) {
+    return res.status(statusCode[err.message] ?? 400).json({
       status: "fail",
-      data: "Update error",
+      message: err.message,
     });
   }
 });
@@ -246,6 +293,10 @@ app.post("/api/edit/:id", async (req, res) => {
 app.post("/api/favorite/:id", async (req, res) => {
   try {
     await mongoose.connect(url);
+
+    if (!(await sessionCheck(req.headers.authorization))) {
+      throw new Error("Unauthorized");
+    }
 
     const filter = { _id: req.params.id };
     const update = { isFavorite: req.body.isFavorite };
@@ -256,14 +307,18 @@ app.post("/api/favorite/:id", async (req, res) => {
       { upsert: true }
     );
 
+    if (!updatedFavoriteContact) {
+      throw new Error("Not Found");
+    }
+
     return res.status(200).json({
       status: "success",
       data: updatedFavoriteContact,
     });
-  } catch {
-    return res.status(400).json({
+  } catch (err) {
+    return res.status(statusCode[err.message] ?? 400).json({
       status: "fail",
-      data: "Update favorite error",
+      message: err.message,
     });
   }
 });
@@ -272,19 +327,37 @@ app.delete("/api/delete/:id", async (req, res) => {
   try {
     await mongoose.connect(url);
 
+    if (!(await sessionCheck(req.headers.authorization))) {
+      throw new Error("Unauthorized");
+    }
+
     const deleted = await Contact.findByIdAndDelete(req.params.id);
+
+    if (!deleted) {
+      throw new Error("Not Found");
+    }
 
     return res.status(200).json({
       status: "success",
       data: deleted,
     });
   } catch (err) {
-    return res.status(400).json({
+    return res.status(statusCode[err.message] ?? 400).json({
       status: "fail",
-      data: "Delete error",
-      error: err.message,
+      message: err.message,
     });
   }
 });
 
 app.listen(3000, () => console.log("blog server running on port 3000!"));
+
+function sessionIDGenerator() {
+  return Math.random() + "" + +new Date();
+}
+
+function sessionCheck(session) {
+  if (!!session) {
+    return User.findOne({ sessionID: session }).then((user) => user);
+  }
+  return Promise.resolve(null);
+}
